Project3/requirements.txt
aiofiles==24.1.0
aiogram==2.25.1
aiohappyeyeballs==2.4.4
aiohttp==3.8.6
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.3.0
Babel==2.9.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.3
frozenlist==1.5.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
kaleido==0.2.1
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.1.0
nest-asyncio==1.6.0
packaging==24.2
plotly==5.24.1
propcache==0.2.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
retrying==1.3.4
six==1.17.0
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.6
yarl==1.18.3
zipp==3.21.0
